chore(payment): preparePayment 디버그 로그 제거 (검증용 임시 로그)

+++ payment-service/src/main/java/com/rushcrew/payment_service/presentation/PaymentPageController.java
@@ -29,7 +29,7 @@ public class PaymentPageController {
             Model model
     ) {
         OrderResponse order = orderClient.getOrder(orderId).data();
-        BigDecimal amount = BigDecimal.valueOf(order.totalAmount());
+        BigDecimal amount = order.totalAmount();
 
         model.addAttribute("orderId", orderId);
         model.addAttribute("amount", amount);
